% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_off.R
\name{cut_off}
\alias{cut_off}
\title{Calculates the cutoff value of ELISA}
\usage{
cut_off(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Calculates the cutoff value of ELISA
}
\references{
Frey, A., Canzio, JD., Zurakowski. 1998. A statistically defined endpoint titer determination method for immunoassays. Journal of Immunological Methods. 221:35-41
}
\author{
Shantanu Tamuly (drsan100@gmail.com)
}
